Doing install of go tools and glide only if they don't exist

+++ install.sh
@@ -1,15 +1,23 @@
 #!/bin/bash
+
+# Install go toolchain
+GO_BINARY="/usr/bin/go"
 GO_VERSION="${GO_VERSION:-1.6}"
 GO_BASE_PATH="/usr/local"
 GOROOT="/usr/local/go"
-rm -rf ${GOROOT}
-curl -sSL https://storage.googleapis.com/golang/go${GO_VERSION}.linux-amd64.tar.gz | tar -xzf - -C ${GO_BASE_PATH}
-ln -s ${GO_BASE_PATH}/go/bin/go /usr/bin/
-ln -s ${GO_BASE_PATH}/go/bin/gofmt /usr/bin/
-ln -s ${GO_BASE_PATH}/go/bin/godoc /usr/bin/
+if [ ! -f ${GO_BINARY} ]; then
+  rm -rf ${GOROOT}
+  curl -sSL https://storage.googleapis.com/golang/go${GO_VERSION}.linux-amd64.tar.gz | tar -xzf - -C ${GO_BASE_PATH}
+  for BINARY in `ls ${GO_BASE_PATH}/go/bin`
+  do
+    ln -s $BINARY /usr/bin/
+  done
+fi
 
 # Install Glide dependency mgmt tool
-curl -sSL https://github.com/Masterminds/glide/releases/download/0.9.1/glide-0.9.1-linux-amd64.tar.gz | tar -xzf - -C /tmp/
-chmod +x /tmp/linux-amd64/glide
-sudo cp /tmp/linux-amd64/glide /usr/bin/
-
+GLIDE_BINARY="/usr/bin/glide"
+if [ ! -f ${GLIDE_BINARY} ]; then
+  curl -sSL https://github.com/Masterminds/glide/releases/download/0.9.1/glide-0.9.1-linux-amd64.tar.gz | tar -xzf - -C /tmp/
+  chmod +x /tmp/linux-amd64/glide
+  mv /tmp/linux-amd64/glide /usr/bin/
+fi
